# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Plantonize BFF API",
    "version": "1.0.0",
    "description": "Documentação completa da API do BFF do Plantonize - proxy e agregação entre frontend Angular, microserviços (Plantões e Notas Fiscais) e Azure Function."
  },
  "servers": [
    {
      "url": "https://plantonize-bff.azurewebsites.net",
      "description": "Servidor de produção (Azure)"
    },
    {
      "url": "http://localhost:{port}",
      "description": "Servidor local (desenvolvimento)",
      "variables": {
        "port": {
          "default": "3000"
        }
      }
    }
  ],
...
  "paths": {
...
    "/api/notas": {
      "get": {
        "tags": ["Notas Fiscais"],
        "summary": "Listar todas as notas fiscais",
        "responses": {"200": {"description": "Lista de notas fiscais", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/Nota"}}}}}}
      },
      "post": {
        "tags": ["Notas Fiscais"],
        "summary": "Criar nota fiscal",
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Nota"}}}},
        "responses": {"201": {"description": "Nota criada"}}
      }
    },
    "/api/notas/emitir": {
      "post": {
        "tags": ["Notas Fiscais"],
        "summary": "Emitir nota fiscal (busca médico e cria nota)",
        "description": "Endpoint que busca dados do médico no microserviço de plantões e cria a nota fiscal",
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/EmitirNotaPayload"}}}},
        "responses": {"201": {"description": "Nota emitida"}, "404": {"description": "Médico não encontrado"}}
      }
    },
    "/api/notas/medico/{medicoId}": {
      "get": {
        "tags": ["Notas Fiscais"],
        "summary": "Listar notas por médico",
        "parameters": [{"name": "medicoId", "in": "path", "required": true, "schema": {"type": "string"}}],
        "responses": {"200": {"description": "Lista de notas do médico"}}
      }
    },
    "/api/notas/{id}": {
      "get": {
        "tags": ["Notas Fiscais"],
        "summary": "Obter nota por ID",
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string"}}],
        "responses": {"200": {"description": "Detalhes da nota"}}
      },
      "put": {
        "tags": ["Notas Fiscais"],
        "summary": "Atualizar nota",
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string"}}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Nota"}}}},
        "responses": {"200": {"description": "Nota atualizada"}}
      },
      "delete": {
        "tags": ["Notas Fiscais"],
        "summary": "Deletar nota",
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string"}}],
        "responses": {"204": {"description": "Nota deletada"}}
      }
    },
    "/api/notas/{id}/exists": {
      "get": {
        "tags": ["Notas Fiscais"],
        "summary": "Verificar se nota existe",
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string"}}],
        "responses": {"200": {"description": "Resultado da verificação"}}
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Nota": {
        "type": "object",
        "properties": {
          "id": {"type": "string"},
          "numeroNota": {"type": "string"},
          "dataEmissao": {"type": "string", "format": "date-time"},
          "valorTotal": {"type": "number"},
          "status": {"type": "integer"},
          "municipioPrestacao": {"type": "string"},
          "issRetido": {"type": "boolean"},
          "medico": {"type": "object"},
          "tomador": {"type": "object"},
          "servicos": {"type": "array"},
          "enviadoEmail": {"type": "boolean"}
        }
      },
      "EmitirNotaPayload": {
        "type": "object",
        "required": ["medicoId", "numeroNota", "dataEmissao", "valorTotal", "municipioPrestacao", "issRetido", "servicos"],
        "properties": {
          "medicoId": {"type": "string"},
          "numeroNota": {"type": "string"},
          "dataEmissao": {"type": "string", "format": "date-time"},
          "valorTotal": {"type": "number"},
          "municipioPrestacao": {"type": "string"},
          "issRetido": {"type": "boolean"},
          "tomador": {"type": "object"},
          "servicos": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "descricao": {"type": "string"},
                "quantidade": {"type": "number"},
                "valorUnitario": {"type": "number"},
                "aliquotaIss": {"type": "number"}
              }
            }
          }
        }
      },
...
    }
  }
}